requirements.txt
# Telegram Bot Framework
aiogram==3.13.1

# 3x-ui API Client
py3xui==0.5.5

# Environment variables
python-dotenv==1.0.1

# QR Code generation (optional, for future use)
qrcode[pil]==7.4.2

# Async HTTP client (dependency of py3xui)
aiohttp==3.10.11

# Scheduler
apscheduler==3.10.4
